Update poll controllers with test
